Moved Navbar component into a separate file for reusability
